+++ android/app/build.gradle
@@ -25,8 +25,10 @@ if (flutterVersionName == null) {
 
 android {
     namespace "com.example.s_chat"
-    compileSdk flutter.compileSdkVersion
-    ndkVersion flutter.ndkVersion
+//    compileSdk flutter.compileSdkVersion
+    compileSdk 35
+//    ndkVersion flutter.ndkVersion
+    ndkVersion "27.0.12077973"
 
     compileOptions {
         // Flag to enable support for the new language APIs
@@ -48,8 +50,8 @@ android {
         applicationId "com.example.s_chat"
         // You can update the following values to match your application needs.
         // For more information, see: https://docs.flutter.dev/deployment/android#reviewing-the-gradle-build-configuration.
-//        minSdkVersion flutter.minSdkVersion
-        minSdkVersion 21
+        minSdkVersion flutter.minSdkVersion
+//        minSdkVersion 21
         targetSdkVersion flutter.targetSdkVersion
         versionCode flutterVersionCode.toInteger()
         versionName flutterVersionName
@@ -73,5 +75,5 @@ dependencies {
 //    implementation platform("com.google.firebase:firebase-bom:32.8.1")
 //    implementation "com.google.firebase:firebase-analytics"
     implementation 'com.android.support:multidex:1.0.3'
-    coreLibraryDesugaring 'com.android.tools:desugar_jdk_libs:1.2.2'
+    coreLibraryDesugaring 'com.android.tools:desugar_jdk_libs:2.1.4'
 }
